refactor logic and test and rename extractor to crawler.

+++ proto/src/main/java/com/doubleknd26/moving/proto/ReviewOrBuilder.java
@@ -8,51 +8,51 @@ public interface ReviewOrBuilder extends
     com.google.protobuf.MessageOrBuilder {
 
   /**
-   * <code>optional .SourceType sourceType = 1;</code>
+   * <code>.SourceType sourceType = 1;</code>
    */
   int getSourceTypeValue();
   /**
-   * <code>optional .SourceType sourceType = 1;</code>
+   * <code>.SourceType sourceType = 1;</code>
    */
   com.doubleknd26.moving.proto.SourceType getSourceType();
 
   /**
-   * <code>optional string url = 2;</code>
+   * <code>string url = 2;</code>
    */
   java.lang.String getUrl();
   /**
-   * <code>optional string url = 2;</code>
+   * <code>string url = 2;</code>
    */
   com.google.protobuf.ByteString
       getUrlBytes();
 
   /**
-   * <code>optional string title = 3;</code>
+   * <code>string title = 3;</code>
    */
   java.lang.String getTitle();
   /**
-   * <code>optional string title = 3;</code>
+   * <code>string title = 3;</code>
    */
   com.google.protobuf.ByteString
       getTitleBytes();
 
   /**
-   * <code>optional int32 grade = 4;</code>
+   * <code>float grade = 4;</code>
    */
-  int getGrade();
+  float getGrade();
 
   /**
-   * <code>optional string comment = 5;</code>
+   * <code>string comment = 5;</code>
    */
   java.lang.String getComment();
   /**
-   * <code>optional string comment = 5;</code>
+   * <code>string comment = 5;</code>
    */
   com.google.protobuf.ByteString
       getCommentBytes();
 
   /**
-   * <code>optional int64 timestamp = 6;</code>
+   * <code>int64 timestamp = 6;</code>
    */
   long getTimestamp();
 }
